feat: run cargo commands in TerrainCLI
Extracting cargo command execution and custom logging to TerrainCLI under runCargoCommand.

Adding the `cargo schema` command to the Terrain build process and removing the build-schema flag and operation from the `terrain contract:generateClient` command.
